Fixed Not Eating if Dish Already In Front of Them
If an NPC starts wanting a dish and the dish is already in front of
them, it originally wouldn't eat it.

Did some refactoring to move the dish-NPC distance into the NPC script
and had it look for a dish when it first gets a desire if the susan is
currently stopped.